extract providers to internal class
